requirements.txt
alabaster==0.7.16
asgiref==3.7.2
Babel==2.14.0
backports.entry-points-selectable==1.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.39.0
colorama==0.4.6
coverage==6.5.0
coveralls==3.3.1
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj3-cloudinary-storage==0.0.6
Django==5.0.2
django-3-jet==1.0.8
django-anymail==10.2
django-appconf==1.0.6
django-autoslug==1.9.9
django-avatar==8.0.0
django-cloudinary-storage==0.3.0
django-countries==7.5.1
django-crispy-forms==2.1
django-embed-video==1.4.9
django-etc==1.4.0
django-extensions==3.2.3
django-gamma-cloudinary==0.2.3
django-grappelli==3.0.8
django-hashid-field==3.4.0
django-hitcount==1.3.5
django-imagekit==5.0.0
django-import-export==3.3.7
django-markdownx==4.0.7
django-markitup==4.1.0
django-next-prev==1.1.0
django-recaptcha==4.0.0
django-robohash-svg==0.9.5
django-sendgrid-v5==1.2.3
django-simple-history==3.5.0
django-slugify-processor==1.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
django-thumbs-v2==0.4.1
django-tinymce==3.7.1
django-tinymce4-lite==1.8.0
django-webtest==1.9.11
djangorestframework==3.14.0
dnspython==2.6.1
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
feedparser==6.0.11
filelock==3.13.1
flake8==7.0.0
future==1.0.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
hashids==1.3.1
httplib2==0.22.0
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
invoke==2.2.0
isort==5.13.2
Jinja2==3.1.3
jsmin==3.0.1
ldif3==3.2.2
Markdown==3.5.2
MarkupPy==1.14
MarkupSafe==2.1.5
mccabe==0.7.0
mock==5.1.0
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pbr==6.0.0
pilkit==3.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.1
python-card-me==0.9.3
python-dateutil==2.8.2
python-http-client==3.3.7
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.11.0
sgmllib3k==1.0.0
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.10.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
surlex==0.2.0
tablib==3.5.0
toml==0.10.2
tox==3.28.0
tox-travis==0.13
typing_extensions==4.10.0
tzdata==2024.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
waitress==3.0.0
webencodings==0.5.1
WebOb==1.8.7
WebTest==3.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
